GadgetGalaxies

Documentation for GadgetGalaxies.

GadgetGalaxies.Galaxy

GadgetGalaxies.Galaxy

GadgetGalaxies.Particles

GadgetGalaxies.Snapshot

GadgetGalaxies.Snapshot

GadgetGalaxies.Snapshot

GadgetGalaxies.convert_units!

GadgetGalaxies.convert_units_full

GadgetGalaxies.convert_units_full_mass

GadgetGalaxies.convert_units_full_pos

GadgetGalaxies.convert_units_full_temp

GadgetGalaxies.convert_units_full_vel

GadgetGalaxies.convert_units_physical

GadgetGalaxies.convert_units_physical!

GadgetGalaxies.convert_units_physical_age

GadgetGalaxies.convert_units_physical_mass

GadgetGalaxies.convert_units_physical_mass!

GadgetGalaxies.convert_units_physical_pos

GadgetGalaxies.convert_units_physical_pos!

GadgetGalaxies.convert_units_physical_temp

GadgetGalaxies.convert_units_physical_temp!

GadgetGalaxies.convert_units_physical_vel

GadgetGalaxies.convert_units_physical_vel!

GadgetGalaxies.convert_units_solar_metallicity

GadgetGalaxies.particle_type_id

GadgetGalaxies.read_halo!

GadgetGalaxies.read_redshift

GadgetGalaxies.shift_across_box_border

GadgetGalaxies.Galaxy — Type

Galaxy(snapshot::Snapshot,
       isub::Integer,
       subid::Union{HaloID,Nothing},
       particles::Dict{Symbol,Particles})

Galaxy of a given snapshot (Snapshot) with the zero-based subhalo index isub with Particles. Including subid (containing information on the exact position in the subfind files) makes accessing properties of the galaxy faster than when only having isub.

The particles are accessible via galaxy.particletype or galaxy[:particletype] (see Particles for typical types).

Examples

julia> p = GadgetGalaxies.Particles(:stars, Dict("ID"=>[1, 2], "MASS"=>[1e11, 2e11]));

julia> g = Galaxy(Snapshot("box", 13), 1532, nothing, Dict(:stars=>p));

julia> g.stars === p
true

GadgetGalaxies.Galaxy — Method

Galaxy(snapshot::Snapshot, isub[, get_id=true])

Convenience method for creating a Galaxy with only a snapshot and a subfind id. Will try to read "MSUB" to extract the if get_id is true.

Particles can still be added after initializing the halo by calling for example galaxy.stars = [...].

GadgetGalaxies.Particles — Type

Particles(type::Symbol, properties::Dict{String})

Particles of a type (typically :stars, :dm, :gas, or :bh) with their properties in a Dict. The Dict has String keys, which are set to all uppercase by default, with exception of iM and Zs.

The properties are accessible via particles.prop (case insensitive).

Examples

julia> p = GadgetGalaxies.Particles(:stars, Dict("ID"=>[1, 2], "MASS"=>[1e11, 2e11]));

julia> p.id
2-element Vector{Int64}:
 1
 2

julia> p.pos = [1 2; 3 4; 5 6]
3×2 Matrix{Int64}:
 1  2
 3  4
 5  6

julia> p.properties["POS"] === p.pos
true

GadgetGalaxies.Snapshot — Type

Snapshot(snapbase, subbase)

Returns a Snapshot from a snapbase and a subbase, which can also be nothing. Both can be files or the base part of files ending with .0, .1, etc.

GadgetGalaxies.Snapshot — Method

Snapshot(path_box, snap::Integer; snapbase=true, subbase=true)

Convenience method for creating a Snapshot from the path to a simulated box and a snapshot number. Setting snapbase or subbase to false sets the respective file base to nothing.

The formats used are the following, for snapbase and subbase, respectively:

snapdir_XXX/snap_XXX

groups_XXX/sub_XXX

GadgetGalaxies.Snapshot — Method

Snapshot(; snapbase=nothing, subbase=nothing)

Convenience method for creating a Snapshot where one of the base files does not exist.

GadgetGalaxies.convert_units! — Function

convert_units!(p::Particles, h::SnapshotHeader, units::Symbol=:full)

Converts the units of the particles properties (see Particles) from a snapshot header in-place. Use :full for units as Unitful quantities, :physical for values converted to physical units (kpc, km/s, solar metallicities etc.), or :sim for values in simulation units. The individual metallicities in "Zs" are converted to a one-dimensional Vector of values in solar metallicities in any case.

GadgetGalaxies.convert_units_full — Function

convert_units_physical(vals::AbstractArray{<:Real}, prop::Symbol, h::SnapshotHeader)
convert_units_physical!(vals::AbstractArray{<:Real}, prop::Symbol, h::SnapshotHeader)
convert_units_full(vals::AbstractArray{<:Real}, prop::Symbol, h::SnapshotHeader)

Converts simulation values to the respective physical values, depending on prop, which can take any of the following values: :pos, :vel, :temp, :mass, :age.

Full returns values in Unitful quantities, whereas physical returns the phyiscal value without unit.

GadgetGalaxies.convert_units_full_mass — Function

convert_units_physical_mass(val::Real, h::SnapshotHeader)
convert_units_physical_mass(vals::AbstractArray{<:Real}, h::SnapshotHeader)
convert_units_physical_mass!(vals::AbstractArray{<:Real}, h::SnapshotHeader)
convert_units_full_mass(val::Real, h::SnapshotHeader)
convert_units_full_mass(vals::AbstractArray{<:Real}, h::SnapshotHeader)

Converts masses via $m \to m \times 10^{10} / h_0$ according to the cosmology defined by the snapshot header. Full returns values in Unitful quantities, whereas physical returns the physical value in M⊙.

GadgetGalaxies.convert_units_full_pos — Function

convert_units_physical_pos(val::Real, h::SnapshotHeader)
convert_units_physical_pos(vals::AbstractArray{<:Real}, h::SnapshotHeader)
convert_units_physical_pos!(vals::AbstractArray{<:Real}, h::SnapshotHeader)
convert_units_full_pos(val::Real, h::SnapshotHeader)
convert_units_full_pos(vals::AbstractArray{<:Real}, h::SnapshotHeader)

Converts positions via $x \to x/(h_0 (z+1)))$ according to the cosmology defined by the snapshot header. Full returns values in Unitful quantities, whereas physical returns the physical value in kpc.

GadgetGalaxies.convert_units_full_temp — Function

convert_units_physical_temp(val::Real, h::SnapshotHeader)
convert_units_physical_temp(vals::AbstractArray{<:Real}, h::SnapshotHeader)
convert_units_physical_temp!(vals::AbstractArray{<:Real}, h::SnapshotHeader)
convert_units_full_temp(val::Real, h::SnapshotHeader)
convert_units_full_temp(vals::AbstractArray{<:Real}, h::SnapshotHeader)

Converts temperatures via $T \to T$ according to the cosmology defined by the snapshot header. Full returns values in Unitful quantities, whereas physical returns the physical value in K.

GadgetGalaxies.convert_units_full_vel — Function

convert_units_physical_vel(val::Real, h::SnapshotHeader)
convert_units_physical_vel(vals::AbstractArray{<:Real}, h::SnapshotHeader)
convert_units_physical_vel!(vals::AbstractArray{<:Real}, h::SnapshotHeader)
convert_units_full_vel(val::Real, h::SnapshotHeader)
convert_units_full_vel(vals::AbstractArray{<:Real}, h::SnapshotHeader)

Converts velocities via $v \to v/\sqrt(z+1))$ according to the cosmology defined by the snapshot header. Full returns values in Unitful quantities, whereas physical returns the physical value in km s^-1.

GadgetGalaxies.convert_units_physical — Function

convert_units_physical(vals::AbstractArray{<:Real}, prop::Symbol, h::SnapshotHeader)
convert_units_physical!(vals::AbstractArray{<:Real}, prop::Symbol, h::SnapshotHeader)
convert_units_full(vals::AbstractArray{<:Real}, prop::Symbol, h::SnapshotHeader)

Converts simulation values to the respective physical values, depending on prop, which can take any of the following values: :pos, :vel, :temp, :mass, :age.

Full returns values in Unitful quantities, whereas physical returns the phyiscal value without unit.

GadgetGalaxies.convert_units_physical! — Function

convert_units_physical(vals::AbstractArray{<:Real}, prop::Symbol, h::SnapshotHeader)
convert_units_physical!(vals::AbstractArray{<:Real}, prop::Symbol, h::SnapshotHeader)
convert_units_full(vals::AbstractArray{<:Real}, prop::Symbol, h::SnapshotHeader)

Converts simulation values to the respective physical values, depending on prop, which can take any of the following values: :pos, :vel, :temp, :mass, :age.

Full returns values in Unitful quantities, whereas physical returns the phyiscal value without unit.

GadgetGalaxies.convert_units_physical_age — Method

convert_units_physical_age(val::Real, h::SnapshotHeader)
convert_units_physical_age(vals::AbstractArray{<:Real}, h::SnapshotHeader)
convert_units_physical_age!(vals::AbstractArray{<:Real}, h::SnapshotHeader)
convert_units_full_age(val::Real, h::SnapshotHeader)
convert_units_full_age(vals::AbstractArray{<:Real}, h::SnapshotHeader)

Converts ages from the scale factor $a$ via the lookback time according to the cosmology defined by the snapshot header. Full returns values in Unitful quantities, whereas physical returns the physical value in Gyr.

GadgetGalaxies.convert_units_physical_mass — Function

convert_units_physical_mass(val::Real, h::SnapshotHeader)
convert_units_physical_mass(vals::AbstractArray{<:Real}, h::SnapshotHeader)
convert_units_physical_mass!(vals::AbstractArray{<:Real}, h::SnapshotHeader)
convert_units_full_mass(val::Real, h::SnapshotHeader)
convert_units_full_mass(vals::AbstractArray{<:Real}, h::SnapshotHeader)

Converts masses via $m \to m \times 10^{10} / h_0$ according to the cosmology defined by the snapshot header. Full returns values in Unitful quantities, whereas physical returns the physical value in M⊙.

GadgetGalaxies.convert_units_physical_mass! — Function

convert_units_physical_mass(val::Real, h::SnapshotHeader)
convert_units_physical_mass(vals::AbstractArray{<:Real}, h::SnapshotHeader)
convert_units_physical_mass!(vals::AbstractArray{<:Real}, h::SnapshotHeader)
convert_units_full_mass(val::Real, h::SnapshotHeader)
convert_units_full_mass(vals::AbstractArray{<:Real}, h::SnapshotHeader)

Converts masses via $m \to m \times 10^{10} / h_0$ according to the cosmology defined by the snapshot header. Full returns values in Unitful quantities, whereas physical returns the physical value in M⊙.

GadgetGalaxies.convert_units_physical_pos — Function

convert_units_physical_pos(val::Real, h::SnapshotHeader)
convert_units_physical_pos(vals::AbstractArray{<:Real}, h::SnapshotHeader)
convert_units_physical_pos!(vals::AbstractArray{<:Real}, h::SnapshotHeader)
convert_units_full_pos(val::Real, h::SnapshotHeader)
convert_units_full_pos(vals::AbstractArray{<:Real}, h::SnapshotHeader)

Converts positions via $x \to x/(h_0 (z+1)))$ according to the cosmology defined by the snapshot header. Full returns values in Unitful quantities, whereas physical returns the physical value in kpc.

GadgetGalaxies.convert_units_physical_pos! — Function

convert_units_physical_pos(val::Real, h::SnapshotHeader)
convert_units_physical_pos(vals::AbstractArray{<:Real}, h::SnapshotHeader)
convert_units_physical_pos!(vals::AbstractArray{<:Real}, h::SnapshotHeader)
convert_units_full_pos(val::Real, h::SnapshotHeader)
convert_units_full_pos(vals::AbstractArray{<:Real}, h::SnapshotHeader)

Converts positions via $x \to x/(h_0 (z+1)))$ according to the cosmology defined by the snapshot header. Full returns values in Unitful quantities, whereas physical returns the physical value in kpc.

GadgetGalaxies.convert_units_physical_temp — Function

convert_units_physical_temp(val::Real, h::SnapshotHeader)
convert_units_physical_temp(vals::AbstractArray{<:Real}, h::SnapshotHeader)
convert_units_physical_temp!(vals::AbstractArray{<:Real}, h::SnapshotHeader)
convert_units_full_temp(val::Real, h::SnapshotHeader)
convert_units_full_temp(vals::AbstractArray{<:Real}, h::SnapshotHeader)

Converts temperatures via $T \to T$ according to the cosmology defined by the snapshot header. Full returns values in Unitful quantities, whereas physical returns the physical value in K.

GadgetGalaxies.convert_units_physical_temp! — Function

convert_units_physical_temp(val::Real, h::SnapshotHeader)
convert_units_physical_temp(vals::AbstractArray{<:Real}, h::SnapshotHeader)
convert_units_physical_temp!(vals::AbstractArray{<:Real}, h::SnapshotHeader)
convert_units_full_temp(val::Real, h::SnapshotHeader)
convert_units_full_temp(vals::AbstractArray{<:Real}, h::SnapshotHeader)

Converts temperatures via $T \to T$ according to the cosmology defined by the snapshot header. Full returns values in Unitful quantities, whereas physical returns the physical value in K.

GadgetGalaxies.convert_units_physical_vel — Function

convert_units_physical_vel(val::Real, h::SnapshotHeader)
convert_units_physical_vel(vals::AbstractArray{<:Real}, h::SnapshotHeader)
convert_units_physical_vel!(vals::AbstractArray{<:Real}, h::SnapshotHeader)
convert_units_full_vel(val::Real, h::SnapshotHeader)
convert_units_full_vel(vals::AbstractArray{<:Real}, h::SnapshotHeader)

Converts velocities via $v \to v/\sqrt(z+1))$ according to the cosmology defined by the snapshot header. Full returns values in Unitful quantities, whereas physical returns the physical value in km s^-1.

GadgetGalaxies.convert_units_physical_vel! — Function

convert_units_physical_vel(val::Real, h::SnapshotHeader)
convert_units_physical_vel(vals::AbstractArray{<:Real}, h::SnapshotHeader)
convert_units_physical_vel!(vals::AbstractArray{<:Real}, h::SnapshotHeader)
convert_units_full_vel(val::Real, h::SnapshotHeader)
convert_units_full_vel(vals::AbstractArray{<:Real}, h::SnapshotHeader)

Converts velocities via $v \to v/\sqrt(z+1))$ according to the cosmology defined by the snapshot header. Full returns values in Unitful quantities, whereas physical returns the physical value in km s^-1.

GadgetGalaxies.convert_units_solar_metallicity — Method

convert_units_solar_metallicity(zs::AbstractMatrix{<:Number}, mass::AbstractVector{<:Number})

Converts an 11×n mass matrix and an n-length mass vector to a metallicity vector of length n, in solar metallicities.

GadgetGalaxies.particle_type_id — Method

particle_type_id(type::Symbol)

Returns the Gadget particle type from a particle Symbol (currently :gas, :dm, :stars, and :bh).

GadgetGalaxies.read_halo! — Method

read_halo!(g::Galaxy; kwargs...)

Reads particles of a Galaxy.

Keywords

radius::Union{Real,Nothing}=nothing: if nothing, the half-mass radius "RHMS"

rad_scale::Real=1: read particles within rad_scale * radius of halo position

units::Symbol=:full: :full for units as Unitful quantities, :physical for values converted to physical units (kpc, km/s, solar metallicities etc.), :sim for values in simulation units

use_keys::Bool=true: if Peano-Hilbert keys should be used if possible

verbose::Bool=false: verbose output

props::Tuple=( (:gas, ["POS", "VEL", "MASS", "TEMP", "SFR", "Zs"]), (:dm, ["POS", "VEL"]), (:stars, ["POS", "VEL", "MASS", "AGE", "Zs", "iM"]), ): particle types (see Particles) and properties to be read

GadgetGalaxies.read_redshift — Method

read_redshift(snapshot::Snapshot)

Returns the redshift z from the first snapfile's header.

GadgetGalaxies.shift_across_box_border — Method

shift_across_box_border(x::Number, x₀::Number, boxsize::Number, boxsize_half::Number)

Shift coordinate x across the box border if reference coordinate x₀ is on the other side.