GadgetGalaxies.Ellipsoid

GadgetGalaxies.Galaxy

GadgetGalaxies.Galaxy

GadgetGalaxies.GalaxyGroup

GadgetGalaxies.GalaxyGroup

GadgetGalaxies.Particles

GadgetGalaxies.Snapshot

GadgetGalaxies.Snapshot

GadgetGalaxies.Snapshot

GadgetGalaxies.Sphere

GadgetGalaxies.angular_momentum

GadgetGalaxies.angular_momentum

GadgetGalaxies.b_value

GadgetGalaxies.b_value

GadgetGalaxies.center_of_mass_iterative

GadgetGalaxies.center_of_mass_iterative

GadgetGalaxies.center_of_mass_iterative

GadgetGalaxies.center_of_velocity

GadgetGalaxies.center_of_velocity

GadgetGalaxies.center_of_velocity

GadgetGalaxies.circular_velocity

GadgetGalaxies.convert_units!

GadgetGalaxies.convert_units_age

GadgetGalaxies.convert_units_full

GadgetGalaxies.convert_units_mass

GadgetGalaxies.convert_units_physical

GadgetGalaxies.convert_units_pos

GadgetGalaxies.convert_units_solar_metallicity

GadgetGalaxies.convert_units_temp

GadgetGalaxies.convert_units_vel

GadgetGalaxies.eccentricity

GadgetGalaxies.ellipsoidal_mask

GadgetGalaxies.ellipticity

GadgetGalaxies.fill_𝐈_up!

GadgetGalaxies.half_mass_radius

GadgetGalaxies.half_mass_radius

GadgetGalaxies.half_mass_radius

GadgetGalaxies.half_mass_radius_2D

GadgetGalaxies.half_mass_radius_2D

GadgetGalaxies.inertia_matrix

GadgetGalaxies.inertia_matrix_iterative

GadgetGalaxies.is_main_halo

GadgetGalaxies.particle_type_id

GadgetGalaxies.read_galaxy_pos

GadgetGalaxies.read_galaxy_prop

GadgetGalaxies.read_galaxy_vel

GadgetGalaxies.read_halo!

GadgetGalaxies.read_header_particle_mass

GadgetGalaxies.read_redshift

GadgetGalaxies.rotate

GadgetGalaxies.rotate

GadgetGalaxies.rotate

GadgetGalaxies.rotate_edgeon

GadgetGalaxies.rotate_edgeon

GadgetGalaxies.rotation_matrix_2D

GadgetGalaxies.rotation_matrix_axis_ratios

GadgetGalaxies.rotation_matrix_edgeon

GadgetGalaxies.rotation_matrix_to_3d

GadgetGalaxies.r²_circle

GadgetGalaxies.r²_ellipsoid

GadgetGalaxies.r²_ellipsoid

GadgetGalaxies.r²_sphere

GadgetGalaxies.shift_across_box_border

GadgetGalaxies.simulation_units_mass

GadgetGalaxies.simulation_units_pos

GadgetGalaxies.simulation_units_temp

GadgetGalaxies.simulation_units_vel

GadgetGalaxies.specific_angular_momentum

GadgetGalaxies.specific_angular_momentum

GadgetGalaxies.spin_parameter

GadgetGalaxies.spin_parameter

GadgetGalaxies.translate

GadgetGalaxies.translate

GadgetGalaxies.translate_to_center_of_mass_iterative

GadgetGalaxies.translate_to_center_of_mass_iterative

GadgetGalaxies.translate_to_center_of_velocity

GadgetGalaxies.translate_to_center_of_velocity

GadgetGalaxies.triaxiality

GadgetGalaxies.Ellipsoid — Type

Ellipsoid(radius::Number, q::U, s::U, constant::Symbol) where U<:Real

Represents an ellipsoid with axis ratios q and s, formed by deforming a sphere of a certain radius at constant :volume or major :axis.

GadgetGalaxies.Galaxy — Type

Galaxy(snapshot::Snapshot,
       isub::Integer,
       subid::Union{HaloID,Nothing},
       particles::Dict{Symbol,Particles})

Galaxy of a given snapshot (Snapshot) with the zero-based subhalo index isub with Particles. Including subid (containing information on the exact position in the subfind files) makes accessing properties of the galaxy faster than when only having isub.

The particles are accessible via galaxy.particletype or galaxy[:particletype] (see Particles for typical types).

Examples

julia> p = GadgetGalaxies.Particles(:stars, Dict("ID"=>[1, 2], "MASS"=>[1e11, 2e11]));

julia> g = Galaxy(Snapshot("box", 13), 1532, nothing, Dict(:stars=>p));

julia> g.stars === p
true

GadgetGalaxies.Galaxy — Method

Galaxy(snapshot::Snapshot, isub[, get_id=true])

Convenience method for creating a Galaxy with only a snapshot and a subfind id. Will try to read "MSUB" to extract the if get_id is true.

Particles can still be added after initializing the galaxy by calling for example galaxy.stars = [...].

GadgetGalaxies.GalaxyGroup — Type

GalaxyGroup(snapshot::Snapshot,
       igroup::Integer,
       groupid::Union{HaloID,Nothing},
       particles::Dict{Symbol,Particles})

Group of a given snapshot (Snapshot) with the zero-based group index igroup with Particles. Including groupid (containing information on the exact position in the subfind files) makes accessing properties of the galaxy faster than when only having igroup.

The particles are accessible via group.particletype or group[:particletype] (see Particles for typical types).

Examples

julia> p = GadgetGalaxies.Particles(:stars, Dict("ID"=>[1, 2], "MASS"=>[1e11, 2e11]));

julia> g = GalaxyGroup(Snapshot("box", 13), 164, nothing, Dict(:stars=>p));

julia> g.stars === p
true

GadgetGalaxies.GalaxyGroup — Method

GalaxyGroup(snapshot::Snapshot, igroup[, get_id=true])

Convenience method for creating a GalaxyGroup with only a snapshot and a subfind id. Will try to read "MTOT" to extract the if get_id is true.

Particles can still be added after initializing the group by calling for example group.stars = [...].

GadgetGalaxies.Particles — Type

Particles(type::Symbol, properties::Dict{String})

Particles of a type (typically :stars, :dm, :gas, or :bh) with their properties in a Dict. The Dict has String keys, which are set to all uppercase by default, with exception of iM and Zs.

The properties are accessible via particles.prop (case insensitive).

Examples

julia> p = GadgetGalaxies.Particles(:stars, Dict("ID"=>[1, 2], "MASS"=>[1e11, 2e11]));

julia> p.id
2-element Vector{Int64}:
 1
 2

julia> p.pos = [1 2; 3 4; 5 6]
3×2 Matrix{Int64}:
 1  2
 3  4
 5  6

julia> p.properties["POS"] === p.pos
true

GadgetGalaxies.Snapshot — Type

Snapshot(snapbase, subbase)

Returns a Snapshot from a snapbase and a subbase, which can also be nothing. Both can be files or the base part of files ending with .0, .1, etc.

GadgetGalaxies.Snapshot — Method

Snapshot(path_box, snap::Integer; snapbase=true, subbase=true)

Convenience method for creating a Snapshot from the path to a simulated box and a snapshot number. Setting snapbase or subbase to false sets the respective file base to nothing.

The formats used are the following, for snapbase and subbase, respectively:

snapdir_XXX/snap_XXX

groups_XXX/sub_XXX

GadgetGalaxies.Snapshot — Method

Snapshot(; snapbase=nothing, subbase=nothing)

Convenience method for creating a Snapshot where one of the base files does not exist.

GadgetGalaxies.Sphere — Method

Sphere(radius::Number)

Returns a spherical Ellipsoid with the specified radius.

GadgetGalaxies.angular_momentum — Method

angular_momentum(pos, vel, mass[, ellipsoid])
angular_momentum(p::Particles[, ellipsoid])

Returns the total angular momentum $J = \sum_i m_i \mathbf{r}_i \times \mathbf{v}_i$ within the ellipsoid if given. Values are assumed to be in kpc, km/s, and solar masses.

using Unitful, UnitfulAstro

g::Galaxy # assume an already read-in galaxy with :full units
sph = Sphere(8u"kpc")

J = angular_momentum(g.stars, sph)

GadgetGalaxies.b_value — Method

b_value(pos, vel, mass[, Mtot])
b_value(pos, vel, mass, ellipsoid)
b_value(p::Particles[, ellipsoid])

Returns the b-value $b = \log(j / (\mathrm{km/s})) - 2/3 \log(M_\mathrm{tot}/\mathrm{M}_\odot)$, where $j$ is the specific_angular_momentum. Values are assumed to be in kpc, km/s, and solar masses.

using Unitful, UnitfulAstro

g::Galaxy # assume an already read-in galaxy with :full units
sph = Sphere(8u"kpc")

b = b_value(g.stars, sph)

GadgetGalaxies.center_of_mass_iterative — Function

center_of_mass_iterative(g::AbstractGalaxy, r_start::Number, ptype::Symbol=:stars; kwargs...)

Returns the center of mass of the galaxy's particles of type ptype (:stars, :dm, :gas, etc.).

GadgetGalaxies.center_of_mass_iterative — Method

center_of_mass_iterative(
    pos::AbstractMatrix{<:Number},
    mass::AbstractVector{<:Number},
    r_start::Number;
    shrinking_factor::Real=0.025,
    limit_fraction::Real=0.01,
    limit_number::Integer=1000,
)

Returns the center of mass by applying the shrinking sphere method (Power et al. 2003).

The sphere starts centered at the coordinate system's origin with a radius of r_start and shrinks by shrinking_factor every iteration. The iteration is stopped when the sphere contains fewer than limit_number and fewer than limit_fraction times the number of particles in the starting sphere, whichever is smaller.

The radius of the sphere in iteration step $i$ is $r_0 (1 - f_\mathrm{shrink})^i$.

GadgetGalaxies.center_of_mass_iterative — Method

center_of_mass_iterative(p::Particles, r_start::Number; kwargs...)

Returns the center of mass of the particles.

GadgetGalaxies.center_of_velocity — Function

center_of_velocity(g::AbstractGalaxy, radius::Number, ptype::Symbol=:stars; kwargs...)

Returns the center of velocity of the galaxy's particles of type ptype (:stars, :dm, :gas, etc.). Requires "POS", "VEL", and "MASS" of the particle type to be available.

GadgetGalaxies.center_of_velocity — Method

center_of_velocity(
    pos::AbstractMatrix{<:Number},
    vel::AbstractMatrix{<:Number},
    mass::AbstractVector{<:Number},
    radius::Number;
    p::Real=0.9,
    r₀::Union{AbstractVector{<:Number},Nothing}=nothing,
)

Returns mass-weighted mean velocity within radius around r₀, only considering absolute velocities smaller than the pth quantile (any value between 0 and 1) when centered around the median velocity.

GadgetGalaxies.center_of_velocity — Method

center_of_velocity(p::Particles, radius::Number; kwargs...)

Returns the center of velocity of the particles. Requires "POS", "VEL", and "MASS" to be available.

GadgetGalaxies.circular_velocity — Method

circular_velocity(Mtot, r)

Returns the circular velocity $v_\mathrm{circ} = \sqrt{G M(<r) / r}$. Values are assumed to be in solar masses and kpc.

GadgetGalaxies.convert_units! — Function

convert_units!(p::Particles, h, units::Symbol=:full)

Converts the units of the particles properties (see Particles) from a snapshot or subfind header in-place. Use :full for units as Unitful quantities, :physical for values converted to physical units (kpc, km/s, solar metallicities etc.), or :sim for values in simulation units. The individual metallicities in "Zs" are converted to a one-dimensional Vector of values in solar metallicities in any case.

GadgetGalaxies.convert_units_age — Function

convert_units_age(val::Union{Real,AbstractArray{<:Real}}, h, units::Symbol=:full)
convert_units_physical_age(val::Real, h)
convert_units_physical_age(vals::AbstractArray{<:Real}, h)
convert_units_physical_age!(vals::AbstractArray{<:Real}, h)
convert_units_full_age(val::Real, h)
convert_units_full_age(vals::AbstractArray{<:Real}, h)

Converts ages from the scale factor $a$ via the lookback time according to the cosmology defined by the snapshot header. Full returns values in Unitful quantities, whereas physical returns the physical value in Gyr.

GadgetGalaxies.convert_units_full — Function

convert_units_physical(vals::AbstractArray{<:Real}, prop::Symbol, h)
convert_units_physical!(vals::AbstractArray{<:Real}, prop::Symbol, h)
convert_units_full(vals::AbstractArray{<:Real}, prop::Symbol, h)

Converts simulation values to the respective physical values, depending on prop, which can take any of the following values: :pos, :vel, :temp, :mass, :age.

Full returns values in Unitful quantities, whereas physical returns the physical value without unit.

GadgetGalaxies.convert_units_mass — Function

convert_units_mass(val::Union{Real,AbstractArray{<:Real}}, h, units::Symbol=:full)
convert_units_physical_mass(val::Real, h)
convert_units_physical_mass(vals::AbstractArray{<:Real}, h)
convert_units_physical_mass!(vals::AbstractArray{<:Real}, h)
convert_units_full_mass(val::Real, h)
convert_units_full_mass(vals::AbstractArray{<:Real}, h)

Converts masses via $m \to m \times 10^{10} / h_0$ according to the cosmology defined by the snapshot header. :full returns values in Unitful quantities, whereas :physical returns the physical value in M⊙. :sim simply returns val.

GadgetGalaxies.convert_units_physical — Function

convert_units_physical(vals::AbstractArray{<:Real}, prop::Symbol, h)
convert_units_physical!(vals::AbstractArray{<:Real}, prop::Symbol, h)
convert_units_full(vals::AbstractArray{<:Real}, prop::Symbol, h)

Converts simulation values to the respective physical values, depending on prop, which can take any of the following values: :pos, :vel, :temp, :mass, :age.

Full returns values in Unitful quantities, whereas physical returns the physical value without unit.

GadgetGalaxies.convert_units_pos — Function

convert_units_pos(val::Union{Real,AbstractArray{<:Real}}, h, units::Symbol=:full)
convert_units_physical_pos(val::Real, h)
convert_units_physical_pos(vals::AbstractArray{<:Real}, h)
convert_units_physical_pos!(vals::AbstractArray{<:Real}, h)
convert_units_full_pos(val::Real, h)
convert_units_full_pos(vals::AbstractArray{<:Real}, h)

Converts positions via $x \to x/(h_0 (z+1)))$ according to the cosmology defined by the snapshot header. :full returns values in Unitful quantities, whereas :physical returns the physical value in kpc. :sim simply returns val.

GadgetGalaxies.convert_units_solar_metallicity — Method

convert_units_solar_metallicity(zs::AbstractMatrix{<:Number}, mass::AbstractVector{<:Number})

Converts an 11×n mass matrix and an n-length mass vector to a metallicity vector of length n, in solar metallicities.

GadgetGalaxies.convert_units_temp — Function

convert_units_temp(val::Union{Real,AbstractArray{<:Real}}, h, units::Symbol=:full)
convert_units_physical_temp(val::Real, h)
convert_units_physical_temp(vals::AbstractArray{<:Real}, h)
convert_units_physical_temp!(vals::AbstractArray{<:Real}, h)
convert_units_full_temp(val::Real, h)
convert_units_full_temp(vals::AbstractArray{<:Real}, h)

Converts temperatures via $T \to T$ according to the cosmology defined by the snapshot header. :full returns values in Unitful quantities, whereas :physical returns the physical value in K. :sim simply returns val.

GadgetGalaxies.convert_units_vel — Function

convert_units_vel(val::Union{Real,AbstractArray{<:Real}}, h, units::Symbol=:full)
convert_units_physical_vel(val::Real, h)
convert_units_physical_vel(vals::AbstractArray{<:Real}, h)
convert_units_physical_vel!(vals::AbstractArray{<:Real}, h)
convert_units_full_vel(val::Real, h)
convert_units_full_vel(vals::AbstractArray{<:Real}, h)

Converts velocities via $v \to v/\sqrt{z+1})$ according to the cosmology defined by the snapshot header. :full returns values in Unitful quantities, whereas :physical returns the physical value in km s^-1. :sim simply returns val.

GadgetGalaxies.eccentricity — Method

eccentricity(q::Real)

Eccentricity from axis ratio q: $e = \sqrt{1 - q^2}$

GadgetGalaxies.ellipsoidal_mask — Method

ellipsoidal_mask(
    pos::AbstractMatrix{<:Number},
    ellipsoid::Ellipsoid;
    return_r²_ellipsoid::Bool=false,
)

Create a BitArray for masking the particles within an ellipsoid.

The ellipsoid is centered around the origin of the coordinate system with its axes placed on the x, y, and z axes, from smallest to largest. If return_r²_ellipsoid is true, returns a tuple of the mask and the vector r² for all particles.

GadgetGalaxies.ellipticity — Method

ellipticity(q::Real)

Ellipticity from axis ratio q: $ϵ = 1 - q$

GadgetGalaxies.fill_𝐈_up! — Method

fill_𝐈_up!(
    𝐈_up::AbstractMatrix{<:Real},
    pos::AbstractMatrix{<:Number},
    mass::Union{AbstractVector{<:Number},Nothing},
    mass_weighted::Bool,
    inertia_matrix_type::Symbol;
    r²::Union{AbstractVector{<:Number},Nothing}=nothing,
)

Fill upper triangle of the inertia matrix for inertia_matrix and inertia_matrix_iterative

Arguments

mass_weighted::Bool: true if positions should be weighted by mass

inertia_matrix_type::Symbol: set to :unweighted or :reduced

Keyword Arguments

r²::Union{AbstractVector{<:Number},Nothing}: if precomputed, the absolute squared distances from the coordinate system's origin; only used when :reduced is used

Note that in the mass-weighted cases, it is not normalized to the total mass, i.e. different than in Eq2/Eq3 of T2015.

GadgetGalaxies.half_mass_radius — Method

half_mass_radius(
    pos::AbstractMatrix{<:Number},
    mass::Union{AbstractVector{<:Number},Nothing},
    [r²::AbstractVector{<:Number}];
    rmax::Union{Number,Nothing}=nothing,
)

Returns the half-mass radius of the particles within radius rmax, or of all particles if rmax is nothing.

GadgetGalaxies.half_mass_radius — Method

half_mass_radius(pos::AbstractMatrix{<:Number}; rmax::Union{Number,Nothing}=nothing)

Returns the half-mass radius of the particles within radius rmax assuming equal masses of all particles, or of all particles if rmax is nothing.

GadgetGalaxies.half_mass_radius — Method

half_mass_radius(g::AbstractGalaxy, ptype::Symbol; kwargs...)
half_mass_radius(p::Particles; kwargs...)

Returns the particles' half-mass radius, depending on if a mass vector is available or not. If the mass is not available, equal particle masses are assumed.

GadgetGalaxies.half_mass_radius_2D — Method

half_mass_radius_2D(
    pos::AbstractMatrix{<:Number},
    [mass::Union{AbstractVector{<:Number},Nothing}];
    rmax::Union{Number,Nothing}=nothing,
    perspective::Symbol=:edgeon,
)

Returns the 2D half-mass radius of the particles within radius rmax when viewed from a given perspective (:edgeon, :sideon, :faceon), or of all particles if rmax is nothing.

GadgetGalaxies.half_mass_radius_2D — Method

half_mass_radius_2D(g::AbstractGalaxy, ptype::Symbol; kwargs...)
half_mass_radius_2D(p::Particles; kwargs...)

Returns the particles' 2D half-mass radius, depending on if a mass vector is available or not. If the mass is not available, equal particle masses are assumed.

GadgetGalaxies.inertia_matrix — Method

inertia_matrix(
    pos::AbstractMatrix{<:Number},
    mass::Union{AbstractVector{<:Number},Nothing};
    mass_weighted::Bool=true,
    inertia_matrix_type::Symbol=:unweighted,
)

Create inertia matrix of different types.

Keyword Arguments

mass_weighted::Bool: true if positions should be weighted by mass

inertia_matrix_type::Symbol: set to :unweighted or :reduced

Inertia matrix types

Unweighted: Eq. 1 in Joachimi et al. 2013,

Mass-weighted unweighted: Eq. 2 in Tenneti et al. 2015,

Reduced: Eq. 3 in Tenneti et al. 2015, but without the mass weighting,

Mass-weighted reduced: Eq. 3 in Tenneti et al. 2015.

Note that in the mass-weighted cases, it is not normalized to the total mass, i.e. different than in Eq2/Eq3 of T2015.

GadgetGalaxies.inertia_matrix_iterative — Method

inertia_matrix_iterative(
    pos::AbstractMatrix{<:Number},
    mass::Union{AbstractVector{<:Number},Nothing},
    inertia_matrix_type::Symbol;
    radius::Union{Number,Nothing}=nothing,
    mass_weighted::Bool=true,
    ellipsoidal_distance::Bool=false,
    constant_volume::Bool=true,
    ϵ::Real=1e-4,
)

Create inertia matrix of different types iteratively (generally more stable than inertia_matrix).

Keyword Arguments

radius::Union{Number,Nothing}: radius of the sphere to start the iteration with

mass_weighted::Bool: true if positions should be weighted by mass

ellipsoidal_distance::Bool: only in combination with reduced inertia matrix type; true to use distances corrected by ellipsoidal axes

constant_volume::Bool: true if the sphere should be deformed into an ellipsoid with constant volume, false for constant major axis

ϵ::Real: iterate until a relative change of the shape axis ratios less than ϵ is encountered

Inertia matrix types

Unweighted: Eq. 1 in Joachimi et al. 2013,

Mass-weighted unweighted: Eq. 2 in Tenneti et al. 2015,

Reduced: Eq. 3 in Tenneti et al. 2015, but without the mass weighting,

Mass-weighted reduced: Eq. 3 in Tenneti et al. 2015.

Reduced with ellipsoidal distances: TODO

Mass-weighted reduced with ellipsoidal distances: TODO

Note that in the mass-weighted cases, it is not normalized to the total mass, i.e. different than in Eq2/Eq3 of T2015.

GadgetGalaxies.is_main_halo — Method

is_main_halo(g::Galaxy)

Returns if the subhalo is the first subhalo in its respective group.

GadgetGalaxies.particle_type_id — Method

particle_type_id(type::Symbol)

Returns the Gadget particle type from a particle Symbol (currently :gas, :dm, :stars, and :bh).

GadgetGalaxies.read_galaxy_pos — Function

read_galaxy_pos(g::AbstractGalaxy, units::Symbol=:full; verbose::Bool=false)

Returns the galaxy's position from subfind using GPOS for a group and SPOS for a subhalo.

GadgetGalaxies.read_galaxy_prop — Function

read_galaxy_prop(g::AbstractGalaxy, prop::AbstractString, units::Symbol=:full; verbose::Bool=false)

Reads galaxy property from subfind for g. Use :full for units as Unitful quantities, :physical for values converted, and :sim for values in simulation units.

GadgetGalaxies.read_galaxy_vel — Function

read_galaxy_vel(g::Galaxy, units::Symbol=:full; verbose::Bool=false)

Returns the galaxy's velocity from subfind using SVEL of FSUB for a group and SVEL for a subhalo.

GadgetGalaxies.read_halo! — Method

read_halo!(g::Galaxy; kwargs...)

Reads particles of a Galaxy.

Keywords

radius::Union{Number,Nothing}=nothing: if nothing, the half-mass radius "RHMS"

radius_units::Symbol=:physical: :physical or :sim, depending on the units of radius

rad_scale::Real=1: read particles within rad_scale * radius of halo position

units::Symbol=:full: use :full for units as Unitful quantities, :physical for values converted to physical units (kpc, km/s, solar metallicities etc.), :sim for values in simulation units

use_keys::Bool=true: if Peano-Hilbert keys should be used if possible

verbose::Bool=false: verbose output

props::Tuple=( (:gas, ["POS", "VEL", "MASS", "TEMP", "SFR", "Zs"]), (:dm, ["POS", "VEL"]), (:stars, ["POS", "VEL", "MASS", "AGE", "Zs", "iM"]) ): particle types (see Particles) and properties to be read

GadgetGalaxies.read_header_particle_mass — Function

read_dm_particle_mass(snapshot::Snapshot, ptype::Symbol[, units=:full]; verbose::Bool=false)

Returns the particle mass from the first snapfile's header for a particle type (:dm, :gas, etc.).

GadgetGalaxies.read_redshift — Method

read_redshift(snapshot::Snapshot)

Returns the redshift $z$ from the first subfile's header.

GadgetGalaxies.rotate — Method

rotate(vals::AbstractMatrix{<:Number}, rotmat::AbstractMatrix{<:Real})
rotate!(vals::AbstractMatrix{<:Number}, rotmat::AbstractMatrix{<:Real})

Returns vals rotated by the rotation matrix rotmat. Works for any dimensions.

GadgetGalaxies.rotate — Method

rotate(g::AbstractGalaxy, rotmat::AbstractMatrix{<:Real})
rotate!(g::AbstractGalaxy, rotmat::AbstractMatrix{<:Real})

Rotates the galaxy g by the $3 \times 3$ rotation matrix rotmat.

The non-in-place version rotate creates a copy of the galaxy with only new pointers to the rotated quantites (generally "POS" and "VEL") and copies of the particle Dicts.

GadgetGalaxies.rotate — Method

rotate(p::Particles, rotmat::AbstractMatrix{<:Real})
rotate!(p::Particles, rotmat::AbstractMatrix{<:Real})

Rotates the particles p by the $3 \times 3$ rotation matrix rotmat.

The non-in-place version rotate creates a copy of the particles with only new pointers to the rotated properties (generally "POS" and "VEL").

GadgetGalaxies.rotate_edgeon — Function

rotate_edgeon(g::AbstractGalaxy, ptype::Symbol=:stars; axis_ratios::Bool=false, kwargs...)
rotate_edgeon!(g::AbstractGalaxy, ptype::Symbol=:stars; axis_ratios::Bool=false, kwargs...)

Rotates the galaxy's particles edge-on for the given particle type ptype (:stars, :dm, :gas, etc.) using the given shape determination algorithm and radius (see rotation_matrix_edgeon for all keyword parameters).

Keyword parameters

axis_ratios::Bool: true to return tuple (g, q, s) with galaxy and axis ratios

GadgetGalaxies.rotate_edgeon — Method

rotate_edgeon(p::Particles; axis_ratios::Bool=false, kwargs...)
rotate_edgeon!(p::Particles; axis_ratios::Bool=false, kwargs...)

Rotates the particles edge-on using the given shape determination algorithm and radius (see rotation_matrix_edgeon for all keyword parameters).

Keyword parameters

axis_ratios::Bool: true to return tuple (p, q, s) with particles and axis ratios

GadgetGalaxies.rotation_matrix_2D — Method

rotation_matrix_2D(
    p::Particles;
    radius::Union{Number,Nothing}=nothing,
    mass_weighted::Bool=true,
    inertia_matrix_type=:unweighted,
    iterative::Bool=true,
    elliptical_distance::Bool=false,
    constant_area::Bool=true,
    algorithm::Union{Symbol,Nothing}=nothing,
    perspective::Symbol=:edgeon,
    matrix_2d::Bool=false,
)

Returns a tuple of the 3D or 2D rotation matrix for rotating the particles in 2D, so that the fitted ellipse is aligned with the axes, and the axis ratio q. If matrix_2d is true, will return the 2D rotation matrix, otherwise the 3D rotation matrix.

See inertia_matrix and inertia_matrix_iterative for the algorithms used for computing the inertia matrix (here only used in 2D, note that terminology is different than in 3D, i.e. elliptical instead of ellipsoidal and constant area instead of volume).

Keyword Arguments

radius::Union{Number,Nothing}: radius of the initial circle for determining the inertia matrix (nothing to include all particles)

mass_weighted::Bool: true if positions should be weighted by mass

inertia_matrix_type: set to :unweighted or :reduced

iterative::Bool: true to use inertia_matrix_iterative, false for intertia_matrix

elliptical_distance::Bool: true only considered together with inertia matrix type :reduced

constant_area::Bool: true if the circle should be deformed into an ellipse with constant area, false for constant major axis

algorithm::Union{Symbol,Nothing}: one of :unw, :red, :unw_i_vol, :unw_i_ax, :red_i_vol, :red_i_ax, :red_ell_i_vol, :red_ell_i_ax (note the same use of vol like in the 3D case instead of area)

perspective::Symbol: one of :edgeon, :sideon, :faceon

matrix_2d::Bool: true to return 2D rotation matrix, false to return only 3D rotation matrix

GadgetGalaxies.rotation_matrix_axis_ratios — Method

rotation_matrix_axis_ratios(𝐈::AbstractMatrix{<:Real})

Returns rotation matrix Q⁻¹ and axis ratios q and s for an inertia matrix 𝐈. s is nothing if 𝐈 is a 2D matrix. If return_axes is true, returns the same tuple with an additional element: an array of the ellipsoid axis vectors.

GadgetGalaxies.rotation_matrix_edgeon — Method

rotation_matrix_edgeon(
    p::Particles;
    radius::Union{Number,Nothing}=nothing,
    mass_weighted::Bool=true,
    inertia_matrix_type=:unweighted,
    iterative::Bool=false,
    ellipsoidal_distance::Bool=false,
    constant_volume::Bool=true,
    algorithm::Union{Symbol,Nothing}=nothing,
    return_axes::Bool=false,
)

Returns a tuple of the 3D rotation matrix for transforming the particles edge-on using the given shape determination algorithm, and the axis ratios q and s. If return_axes is true, returns the same tuple with an additional element: an array of the ellipsoid axis vectors.

See inertia_matrix and inertia_matrix_iterative for the algorithms used for computing the inertia matrix.

Keyword Arguments

radius::Union{Number,Nothing}: radius of the initial sphere for determining the inertia matrix (nothing to include all particles)

mass_weighted::Bool: true if positions should be weighted by mass

inertia_matrix_type: set to :unweighted or :reduced

iterative::Bool: true to use inertia_matrix_iterative, false for intertia_matrix

ellipsoidal_distance::Bool: true only considered together with inertia matrix type :reduced

constant_volume::Bool: true if the sphere should be deformed into an ellipsoid with constant volue, false for constant major axis

algorithm::Union{Symbol,Nothing}: one of :unw, :red, :unw_i_vol, :unw_i_ax, :red_i_vol, :red_i_ax, :red_ell_i_vol, :red_ell_i_ax

return_axes::Bool: true to return tuple (rotmat, q, s), false to return only rotmat

GadgetGalaxies.rotation_matrix_to_3d — Method

rotation_matrix_to_3d(R::AbstractMatrix{T}, dims::AbstractVector{<:Integer}) where T<:Real

Converts a 2D rotation matrix R in the plane given by dimensions dims (e.g. [1, 2] or [3, 1]) to a 3D rotation matrix.

GadgetGalaxies.r²_circle — Method

r²_circle(pos::AbstractMatrix{<:Number})

Squared distances in 2D from origin.

GadgetGalaxies.r²_ellipsoid — Method

r²_ellipsoid(pos::AbstractMatrix{<:Number}, q::Real, s::Nothing)

Squared ellipsoidal distances in 2D from origin with axis ratio q.

For 3D version, see Eq. 6 in Allgood et al. 2006, https://arxiv.org/abs/astro-ph/0508497

GadgetGalaxies.r²_ellipsoid — Method

r²_ellipsoid(pos::AbstractMatrix{<:Number}, q::Real, s::Real)

Squared ellipsoidal distances from origin with axis ratios q and s.

See Eq. 6 in Allgood et al. 2006, https://arxiv.org/abs/astro-ph/0508497

GadgetGalaxies.r²_sphere — Method

r²_sphere(pos::AbstractMatrix{<:Number})

Squared distances from origin in 2D (calls r²_circle in those cases) and 3D.

GadgetGalaxies.shift_across_box_border — Method

shift_across_box_border(x::Number, x₀::Number, boxsize::Number, boxsize_half::Number)

Shift coordinate x across the box border if reference coordinate x₀ is on the other side.

GadgetGalaxies.simulation_units_mass — Function

simulation_units_mass(val::Number, h)
simulation_units_mass(vals::AbstractArray{<:Number}, h)
simulation_units_mass!(vals::AbstractArray{<:Real}, h)

Converts masses via $m \to m \times h_0 / 10^{10}$ according to the cosmology defined by the snapshot header from physical to simulation units. Unitless parameters need to be given in M⊙.

GadgetGalaxies.simulation_units_pos — Function

simulation_units_pos(val::Number, h)
simulation_units_pos(vals::AbstractArray{<:Number}, h)
simulation_units_pos!(vals::AbstractArray{<:Real}, h)

Converts positions via $x \to x \times h_0 (z+1)$ according to the cosmology defined by the snapshot header from physical to simulation units. Unitless parameters need to be given in kpc.

GadgetGalaxies.simulation_units_temp — Function

simulation_units_temp(val::Number, h)
simulation_units_temp(vals::AbstractArray{<:Number}, h)
simulation_units_temp!(vals::AbstractArray{<:Real}, h)

Converts temperatures via $T \to T$ according to the cosmology defined by the snapshot header from physical to simulation units. Unitless parameters need to be given in K.

GadgetGalaxies.simulation_units_vel — Function

simulation_units_vel(val::Number, h)
simulation_units_vel(vals::AbstractArray{<:Number}, h)
simulation_units_vel!(vals::AbstractArray{<:Real}, h)

Converts velocities via $v \to v \times \sqrt{z+1}$ according to the cosmology defined by the snapshot header from physical to simulation units. Unitless parameters need to be given in km s^-1.

GadgetGalaxies.specific_angular_momentum — Method

specific_angular_momentum(pos, vel, mass; Mtot)
specific_angular_momentum(pos, vel, mass, ellipsoid)
specific_angular_momentum(p::Particles[, ellipsoid]; Mtot)

Returns the specific angular momentum $j = \sum_i m_i \mathbf{r}_i \times \mathbf{v}_i / \sum_i m_i$ within the ellipsoid if given. Values are assumed to be in kpc, km/s, and solar masses. The total mass can be provided via Mtot if already known.

using Unitful, UnitfulAstro

g::Galaxy # assume an already read-in galaxy with :full units
sph = Sphere(8u"kpc")

j = specific_angular_momentum(g.stars, sph)

GadgetGalaxies.spin_parameter — Method

spin_parameter(j, Mtot)
spin_parameter(pos, vel, mass, Mtot, r)
spin_parameter(pos, vel, mass, Mtot, sphere)
spin_parameter(p::Particles, Mtot, r; mask=nothing)

Returns the spin parameter $\lambda(r) = j / (\sqrt{2} r v_\mathrm{circ}(r))$, where $j$ is the specific_angular_momentum. The spin parameter is computed for only one particle component.

A mask can be applied to filter the particles.

GadgetGalaxies.spin_parameter — Method

spin_parameter(g::AbstractGalaxy, Mtot, r)

Returns the total spin parameter $\lambda(r) = j / (\sqrt{2} r v_\mathrm{circ}(r))$, where $j$ is the specific_angular_momentum. The spin parameter is computed for all particle components.

GadgetGalaxies.translate — Function

translate(g::AbstractGalaxy, x⃗::AbstractVector{<:Number}, prop::Symbol=:pos)
translate!(g::AbstractGalaxy, x⃗::AbstractVector{<:Number}, prop::Symbol=:pos)

Translates the galaxy g by the vector x⃗.

The non-in-place version translate creates a copy of the galaxy with only new pointers to the property and copies of the particle Dicts.

GadgetGalaxies.translate — Function

translate(p::Particles, x⃗::AbstractVector{<:Number}, prop::Symbol=:pos)
translate!(p::Particles, x⃗::AbstractVector{<:Number}, prop::Symbol=:pos)

Translates the particles p by the vector x⃗.

The non-in-place version translate creates a copy of the particles with only new pointers to the property.

GadgetGalaxies.translate_to_center_of_mass_iterative — Function

translate_to_center_of_mass_iterative(g::AbstractGalaxy, r_start::Number, ptype::Symbol=:stars; kwargs...)
translate_to_center_of_mass_iterative!(g::AbstractGalaxy, r_start::Number, ptype::Symbol=:stars; kwargs...)

Translates the particles to their center of mass of the particles (see center_of_mass_iterative).

GadgetGalaxies.translate_to_center_of_mass_iterative — Method

translate_to_center_of_mass_iterative(p::Particles, r_start::Number; kwargs...)
translate_to_center_of_mass_iterative!(p::Particles, r_start::Number; kwargs...)

Translates the particles to their center of mass of the particles (see center_of_mass_iterative).

GadgetGalaxies.translate_to_center_of_velocity — Function

translate_to_center_of_velocity(g::AbstractGalaxy, radius::Number, ptype::Symbol=:stars; kwargs...)
translate_to_center_of_velocity!(g::AbstractGalaxy, radius::Number, ptype::Symbol=:stars; kwargs...)

Translates the particles to their center of velocity of the particles (see center_of_velocity). Requires "POS", "VEL", and "MASS" of the particle type to be available.

GadgetGalaxies.translate_to_center_of_velocity — Method

translate_to_center_of_velocity(p::Particles, radius::Number; kwargs...)
translate_to_center_of_velocity!(p::Particles, radius::Number; kwargs...)

Translates the particles to their center of velocity of the particles (see center_of_velocity). Requires "POS", "VEL", and "MASS" to be available.

GadgetGalaxies.triaxiality — Method

triaxiality(q::Real, s::Real)

Triaxiality from the axis ratios q and s.

$0 < T < 1/3$: oblate

$1/3 < T < 2/3$: triaxial

$2/3 < T < 1$: prolate