Kinematic Parameters

GadgetGalaxies provides various functions to compute kinematic parameters of galaxies.

Angular Momentum

The (specific) angular momentum can easily be computed from a Galaxy or GalaxyGroup:

GadgetGalaxies.angular_momentum — Function

angular_momentum(pos, vel, mass[, ellipsoid])
angular_momentum(p::Particles[, ellipsoid])

Returns the total angular momentum $J = \sum_i m_i \mathbf{r}_i \times \mathbf{v}_i$ within the ellipsoid if given. Values are assumed to be in kpc, km/s, and solar masses.

using Unitful, UnitfulAstro

g::Galaxy # assume an already read-in galaxy with :full units
sph = Sphere(8u"kpc")

J = angular_momentum(g.stars, sph)

GadgetGalaxies.specific_angular_momentum — Function

specific_angular_momentum(pos, vel, mass; Mtot)
specific_angular_momentum(pos, vel, mass, ellipsoid)
specific_angular_momentum(p::Particles[, ellipsoid]; Mtot)

Returns the specific angular momentum $j = \sum_i m_i \mathbf{r}_i \times \mathbf{v}_i / \sum_i m_i$ within the ellipsoid if given. Values are assumed to be in kpc, km/s, and solar masses. The total mass can be provided via Mtot if already known.

using Unitful, UnitfulAstro

g::Galaxy # assume an already read-in galaxy with :full units
sph = Sphere(8u"kpc")

j = specific_angular_momentum(g.stars, sph)

B-Value

The b-value methods for a galaxy are called in the same way as the angular momenta methods. However, since this is generally only computed for the stellar component, it is recommended to center the stars positionally and kinematically:

using Unitful, UnitfulAstro

gr::GalaxyGroup # assume an already read-in galaxy group with :full units

rvir = read_galaxy_prop(gr, "RVIR")

# first half-mass radius estimate
r = half_mass_radius(gr.stars; rmax=0.1rvir)

# shift to center
translate_to_center_of_mass_iterative!(gr, 4r, :stars)

# second determination of the half-mass radius (this time with proper centering)
r = half_mass_radius(gr.stars; rmax=0.1rvir)

# center velocity
translate_to_center_of_velocity!(gr, 4r)

# finally determine b_value within two half-mass radii
b = b_value(gr.stars, Sphere(2r))

GadgetGalaxies.b_value — Function

b_value(pos, vel, mass[, Mtot])
b_value(pos, vel, mass, ellipsoid)
b_value(p::Particles[, ellipsoid])

Returns the b-value $b = \log(j / (\mathrm{km/s})) - 2/3 \log(M_\mathrm{tot}/\mathrm{M}_\odot)$, where $j$ is the specific_angular_momentum. Values are assumed to be in kpc, km/s, and solar masses.

using Unitful, UnitfulAstro

g::Galaxy # assume an already read-in galaxy with :full units
sph = Sphere(8u"kpc")

b = b_value(g.stars, sph)