GadgetGalaxies

Documentation for GadgetGalaxies.

GadgetGalaxies.Galaxy

GadgetGalaxies.Galaxy

GadgetGalaxies.Particles

GadgetGalaxies.Snapshot

GadgetGalaxies.Snapshot

GadgetGalaxies.Snapshot

GadgetGalaxies.particle_type_id

GadgetGalaxies.shift_across_box_border

GadgetGalaxies.Galaxy — Type

Galaxy(snapshot::Snapshot,
       isub::Integer,
       subid::Union{HaloID,Nothing},
       particles::Dict{Symbol,Particles})

Galaxy of a given snapshot (Snapshot) with the zero-based subhalo index isub with Particles. Including subid (containing information on the exact position in the subfind files) makes accessing properties of the galaxy faster than when only having isub.

The particles are accessible via galaxy.particletype or galaxy[:particletype] (see Particles for typical types).

Examples

julia> p = GadgetGalaxies.Particles(:stars, Dict("ID"=>[1, 2], "MASS"=>[1e11, 2e11]));

julia> g = Galaxy(Snapshot("box", 13), 1532, nothing, Dict(:stars=>p));

julia> g.stars === p
true

GadgetGalaxies.Galaxy — Method

Galaxy(snapshot::Snapshot, isub[, get_id=true])

Convenience method for creating a Galaxy with only a snapshot and a subfind id. Will try to read "MSUB" to extract the HaloID if get_id is true.

Particles can still be added after initializing the halo by calling for example galaxy.stars = [...].

GadgetGalaxies.Particles — Type

Particles(type::Symbol, properties::Dict{String})

Particles of a type (typically :stars, :dm, :gas, or :bh) with their properties in a Dict. The Dict has String keys, which are set to all uppercase by default, with exception of iM and Zs.

The properties are accessible via particles.prop (case insensitive).

Examples

julia> p = GadgetGalaxies.Particles(:stars, Dict("ID"=>[1, 2], "MASS"=>[1e11, 2e11]));

julia> p.id
2-element Vector{Int64}:
 1
 2

julia> p.pos = [1 2; 3 4; 5 6]
3×2 Matrix{Int64}:
 1  2
 3  4
 5  6

julia> p.properties["POS"] === p.pos
true

GadgetGalaxies.Snapshot — Type

Snapshot(snapbase, subbase)

Returns a Snapshot from a snapbase and a subbase, which can also be nothing. Both can be files or the base part of files ending with .0, .1, etc.

GadgetGalaxies.Snapshot — Method

Snapshot(path_box, snap::Integer; snapbase=true, subbase=true)

Convenience method for creating a Snapshot from the path to a simulated box and a snapshot number. Setting snapbase or subbase to false sets the respective file base to nothing.

The formats used are the following, for snapbase and subbase, respectively:

snapdir_XXX/snap_XXX

groups_XXX/sub_XXX

GadgetGalaxies.Snapshot — Method

Snapshot(; snapbase=nothing, subbase=nothing)

Convenience method for creating a Snapshot where one of the base files does not exist.

GadgetGalaxies.particle_type_id — Method

particle_type_id(type::Symbol)

Returns the Gadget particle type from a particle Symbol (currently :gas, :dm, :stars,and bh).

GadgetGalaxies.shift_across_box_border — Method

shift_across_box_border(x::Number, x₀::Number, boxsize::Number, boxsize_half::Number)

Shift coordinate x across the box border if reference coordinate x₀ is on the other side.